bashrc update and docs

+++ .bashrc
@@ -142,9 +142,11 @@ engnotes(){
     cd ~
 }
 venv(){
-    python3 -m venv venv
-    source venv/bin/activate
-    python3 -m pip install -r requirements.txt
+    python3 -m venv $1
+    source $1/bin/activate
+    if [[ -f "./requirements.txt" ]];then
+      python3 -m pip install -r requirements.txt
+    fi
 }
 clean-olm(){
   oc delete -n openshift-marketplace `oc get job -n openshift-marketplace -o name`
